+++ src/ext/java/org/opentripplanner/ext/flex/FlexLegMapper.java
@@ -10,6 +10,14 @@ import org.opentripplanner.routing.algorithm.mapping.GraphPathToItineraryMapper;
 public class FlexLegMapper {
 
     static public void fixFlexTripLeg(Leg leg, FlexTripEdge flexTripEdge) {
+      leg.from.stop = flexTripEdge.s1;
+      // TODO: Should flex be of its own type
+      leg.from.vertexType = flexTripEdge.s1 instanceof Stop ? VertexType.TRANSIT : VertexType.NORMAL;
+      leg.from.stopIndex = flexTripEdge.flexTemplate.fromStopIndex;
+      leg.to.stop = flexTripEdge.s2;
+      leg.to.vertexType = flexTripEdge.s2 instanceof Stop ? VertexType.TRANSIT : VertexType.NORMAL;
+      leg.to.stopIndex = flexTripEdge.flexTemplate.toStopIndex;
+
         leg.intermediateStops = new ArrayList<>();
         leg.distanceMeters = flexTripEdge.getDistanceMeters();