+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/plan/PlanPlaceType.java
@@ -8,13 +8,19 @@ import graphql.schema.GraphQLOutputType;
 import org.opentripplanner.ext.transmodelapi.model.EnumTypes;
 import org.opentripplanner.ext.transmodelapi.model.scalars.GeoJSONCoordinatesScalar;
 import org.opentripplanner.ext.transmodelapi.support.GqlUtil;
+import org.opentripplanner.model.FlexStopLocation;
+import org.opentripplanner.model.Stop;
+import org.opentripplanner.model.StopLocation;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.VertexType;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalVehicle;
 
 public class PlanPlaceType {
 
   public static GraphQLObjectType create(
       GraphQLOutputType bikeRentalStationType,
+      GraphQLOutputType rentalVehicleType,
       GraphQLOutputType quayType
   ) {
     return GraphQLObjectType
@@ -28,7 +34,7 @@ public class PlanPlaceType {
             .description(
                 "For transit quays, the name of the quay. For points of interest, the name of the POI.")
             .type(Scalars.GraphQLString)
-            .dataFetcher(environment -> ((Place) environment.getSource()).getName())
+            .dataFetcher(environment -> ((Place) environment.getSource()).name)
             .build())
         .field(GraphQLFieldDefinition
             .newFieldDefinition()
@@ -36,40 +42,38 @@ public class PlanPlaceType {
             .description(
                 "Type of vertex. (Normal, Bike sharing station, Bike P+R, Transit quay) Mostly used for better localization of bike sharing and P+R station names")
             .type(EnumTypes.VERTEX_TYPE)
-            .dataFetcher(environment -> ((Place) environment.getSource()).getVertexType())
+            .dataFetcher(environment -> ((Place) environment.getSource()).vertexType)
             .build())
         .field(GraphQLFieldDefinition
             .newFieldDefinition()
             .name("latitude")
             .description("The latitude of the place.")
             .type(new GraphQLNonNull(Scalars.GraphQLFloat))
-            .dataFetcher(environment -> ((Place) environment.getSource()).getCoordinate().latitude())
+            .dataFetcher(environment -> ((Place) environment.getSource()).coordinate.latitude())
             .build())
         .field(GraphQLFieldDefinition
             .newFieldDefinition()
             .name("longitude")
             .description("The longitude of the place.")
             .type(new GraphQLNonNull(Scalars.GraphQLFloat))
-            .dataFetcher(environment -> ((Place) environment.getSource()).getCoordinate().longitude())
+            .dataFetcher(environment -> ((Place) environment.getSource()).coordinate.longitude())
             .build())
         .field(GraphQLFieldDefinition
             .newFieldDefinition()
             .name("quay")
             .description("The quay related to the place.")
             .type(quayType)
-            .dataFetcher(environment -> ((Place) environment.getSource()).getStopId() != null
-                ? GqlUtil.getRoutingService(environment)
-                .getStopForId(((Place) environment.getSource()).getStopId()) : null)
+            .dataFetcher(environment -> ((Place) environment.getSource()).stop instanceof Stop
+                ? ((Stop) ((Place) environment.getSource()).stop)
+                : null)
             .build())
         .field(GraphQLFieldDefinition
             .newFieldDefinition()
             .name("flexibleArea")
             .description("The flexible area related to the place.")
             .type(GeoJSONCoordinatesScalar.getGraphQGeoJSONCoordinatesScalar())
-            .dataFetcher(environment -> ((Place) environment.getSource()).getStopId() != null
-                ? GqlUtil.getRoutingService(environment)
-                .getLocationById(((Place) environment.getSource()).getStopId())
-                .getGeometry().getCoordinates()
+            .dataFetcher(environment -> ((Place) environment.getSource()).stop instanceof FlexStopLocation
+                ? ((FlexStopLocation) ((Place) environment.getSource()).stop).getGeometry().getCoordinates()
                 : null)
             .build())
         .field(GraphQLFieldDefinition
@@ -77,7 +81,20 @@ public class PlanPlaceType {
             .name("bikeRentalStation")
             .type(bikeRentalStationType)
             .description("The bike rental station related to the place")
-            .dataFetcher(environment -> ((Place) environment.getSource()).getBikeRentalStation())
+            .dataFetcher(environment -> ((Place) environment.getSource()).vertexType.equals(VertexType.BIKESHARE)
+                    && ((Place) environment.getSource()).vehicleRentalStation instanceof VehicleRentalStation
+                ? ((Place) environment.getSource()).vehicleRentalStation
+                : null)
+            .build())
+        .field(GraphQLFieldDefinition
+            .newFieldDefinition()
+            .name("rentalVehicle")
+            .type(rentalVehicleType)
+            .description("The rental vehicle related to the place")
+            .dataFetcher(environment -> ((Place) environment.getSource()).vertexType.equals(VertexType.BIKESHARE)
+                    && ((Place) environment.getSource()).vehicleRentalStation instanceof VehicleRentalVehicle
+                ? ((Place) environment.getSource()).vehicleRentalStation
+                : null)
             .build())
         //                .field(GraphQLFieldDefinition.newFieldDefinition()
         //                        .name("bikePark")