+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/GraphPathToItineraryMapper.java
@@ -20,6 +20,7 @@ import org.opentripplanner.ext.flex.FlexLegMapper;
 import org.opentripplanner.ext.flex.edgetype.FlexTripEdge;
 import org.opentripplanner.model.VehicleRentalStationInfo;
 import org.opentripplanner.model.StreetNote;
+import org.opentripplanner.model.VehicleRentalStationInfo;
 import org.opentripplanner.model.WgsCoordinate;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.Leg;
@@ -31,21 +32,23 @@ import org.opentripplanner.routing.core.RoutingContext;
 import org.opentripplanner.routing.core.State;
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.edgetype.AreaEdge;
-import org.opentripplanner.routing.edgetype.VehicleRentalEdge;
 import org.opentripplanner.routing.edgetype.ElevatorAlightEdge;
 import org.opentripplanner.routing.edgetype.FreeEdge;
 import org.opentripplanner.routing.edgetype.PathwayEdge;
 import org.opentripplanner.routing.edgetype.StreetEdge;
+import org.opentripplanner.routing.edgetype.VehicleRentalEdge;
 import org.opentripplanner.routing.edgetype.VehicleParkingEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.location.TemporaryStreetLocation;
 import org.opentripplanner.routing.spt.GraphPath;
+import org.opentripplanner.routing.vertextype.BikeParkVertex;
 import org.opentripplanner.routing.vertextype.VehicleRentalStationVertex;
 import org.opentripplanner.routing.vertextype.ExitVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.TransitStopVertex;
+import org.opentripplanner.routing.vertextype.VehicleRentalStationVertex;
 import org.opentripplanner.routing.vertextype.VehicleParkingEntranceVertex;
 import org.opentripplanner.util.OTPFeature;
 import org.opentripplanner.util.PolylineEncoder;
@@ -298,10 +301,6 @@ public abstract class GraphPathToItineraryMapper {
 
         leg.generalizedCost = (int) (states[states.length - 1].getWeight() - states[0].getWeight());
 
-        // Interlining information is now in a separate field in Graph, not in edges.
-        // But in any case, with Raptor this method is only being used to translate non-transit legs of paths.
-        leg.interlineWithPreviousLeg = false;
-
         leg.walkingBike = states[states.length - 1].isBackWalkingBike();
 
         leg.walkingBike = states[states.length - 1].isBackWalkingBike();
@@ -466,8 +465,11 @@ public abstract class GraphPathToItineraryMapper {
      * @param states The states that go with the leg
      */
     private static void addPlaces(Leg leg, State[] states, Locale requestedLocale) {
-        leg.from = makePlace(states[0], requestedLocale);
-        leg.to = makePlace(states[states.length - 1], requestedLocale);
+        Vertex firstVertex = states[0].getVertex();
+        Vertex lastVertex = states[states.length - 1].getVertex();
+
+        leg.from = makePlace(firstVertex, requestedLocale);
+        leg.to = makePlace(lastVertex, requestedLocale);
     }
 
     /**
@@ -477,8 +479,11 @@ public abstract class GraphPathToItineraryMapper {
      * @param requestedLocale The locale to use for all text attributes.
      * @return The resulting {@link Place} object.
      */
-    private static Place makePlace(State state, Locale requestedLocale) {
-        Vertex vertex = state.getVertex();
+    private static Place makePlace(Vertex vertex, Locale requestedLocale) {
+        if (vertex instanceof TransitStopVertex) {
+            return new Place(((TransitStopVertex) vertex).getStop());
+        }
+
         String name = vertex.getName(requestedLocale);
 
         //This gets nicer names instead of osm:node:id when changing mode of transport
@@ -488,6 +493,13 @@ public abstract class GraphPathToItineraryMapper {
             name = ((StreetVertex) vertex).getIntersectionName(requestedLocale).toString(requestedLocale);
         }
 
+
+        if (vertex instanceof VehicleRentalStationVertex) {
+            place.vehicleRentalStation = ((VehicleRentalStationVertex) vertex).getStation();
+            LOG.trace("Added bike share Id {} to place", place.vehicleRentalStation.getId());
+            place.vertexType = VertexType.BIKESHARE;
+        } else if (vertex instanceof BikeParkVertex) {
+            place.vertexType = VertexType.BIKEPARK;
         if (vertex instanceof TransitStopVertex) {
             return Place.forStop((TransitStopVertex) vertex, name);
         } else if(vertex instanceof VehicleRentalStationVertex) {