+++ src/test/java/org/opentripplanner/routing/algorithm/GraphRoutingTest.java
@@ -11,9 +11,13 @@ import org.locationtech.jts.geom.Coordinate;
 import org.opentripplanner.common.TurnRestriction;
 import org.opentripplanner.common.TurnRestrictionType;
 import org.opentripplanner.common.geometry.GeometryUtils;
+import org.opentripplanner.model.Agency;
 import org.opentripplanner.model.Entrance;
 import org.opentripplanner.model.FeedScopedId;
+import org.opentripplanner.model.Route;
 import org.opentripplanner.model.Stop;
+import org.opentripplanner.model.StopTime;
+import org.opentripplanner.model.TransitMode;
 import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.model.WgsCoordinate;
 import org.opentripplanner.model.WheelChairBoarding;
@@ -36,12 +40,13 @@ import org.opentripplanner.routing.edgetype.StreetVehicleParkingLink;
 import org.opentripplanner.routing.edgetype.StreetVehicleRentalLink;
 import org.opentripplanner.routing.edgetype.TemporaryFreeEdge;
 import org.opentripplanner.routing.edgetype.VehicleRentalEdge;
-import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalPlace;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.GraphIndex;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.location.TemporaryStreetLocation;
+import org.opentripplanner.routing.vehicle_rental.VehicleRentalStation;
 import org.opentripplanner.routing.spt.GraphPath;
 import org.opentripplanner.routing.spt.ShortestPathTree;
 import org.opentripplanner.routing.vehicle_parking.VehicleParking;
@@ -300,7 +305,7 @@ public abstract class GraphRoutingTest {
         }
 
         // -- Vehicle rental
-        public VehicleRentalStation vehicleRentalStationEntity(
+        public VehicleRentalPlace vehicleRentalStationEntity(
                 String id,
                 double latitude,
                 double longitude,
@@ -310,6 +315,8 @@ public abstract class GraphRoutingTest {
             vehicleRentalStation.id = new FeedScopedId(network, id);
             vehicleRentalStation.longitude = longitude;
             vehicleRentalStation.latitude = latitude;
+            vehicleRentalStation.vehiclesAvailable = 2;
+            vehicleRentalStation.spacesAvailable = 2;
             vehicleRentalStation.isKeepingVehicleRentalAtDestinationAllowed = false;
             vehicleRentalStation.name = new NonLocalizedString(id);
             return vehicleRentalStation;
@@ -399,9 +406,26 @@ public abstract class GraphRoutingTest {
             return List.of(link(from, to), link(to, from));
         }
 
+        public Agency agency(String name) {
+            return new Agency(new FeedScopedId("Test", name), name, null);
+        }
+
+        public Route route(String id, TransitMode mode, Agency agency) {
+            var route = new Route(new FeedScopedId("Test", id));
+            route.setAgency(agency);
+            route.setMode(mode);
+            return route;
+        }
+
         // Transit
         public void tripPattern(TripPattern tripPattern) {
             graph.tripPatternForId.put(tripPattern.getId(), tripPattern);
+        }
+
+        public StopTime st(TransitStopVertex s1) {
+            var st = new StopTime();
+            st.setStop(s1.getStop());
+            return st;
         }
     }